% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_numeric.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert a character column to a numeric column}
\usage{
convert_to_numeric(df, column_name)
}
\arguments{
\item{df}{A data frame.}

\item{column_name}{The name of the column to convert.}
}
\value{
A data frame with the specified column converted to numeric. Non-numeric values are converted to NA.
}
\description{
Convert a character column to a numeric column
}
\examples{
data(mtcars)
mtcars$mpg <- as.character(mtcars$mpg)
convert_to_numeric(mtcars, "mpg")
}
