% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing.R
\name{remove_missing}
\alias{remove_missing}
\title{Remove rows with missing values from a data frame}
\usage{
remove_missing(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with rows containing missing values removed.
}
\description{
Remove rows with missing values from a data frame
}
\examples{
data(mtcars)
mtcars[1,1] <- NA
remove_missing(mtcars)
}
